Move battery entities to binary_sensor platform
